pyproject.toml
[tool.poetry]
name = "gif"
version = "0.1.0"
description = ""
authors = ["Richie <richiekim501@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
